+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v0.0.2]
+
+### Added
+- New minimal centrifuge database included with the project
+- New minimal sample dataset for testing
+
+### Changed
+- Report reads stats section using standard aplanat report section
+
+### Removed
+- Report section comparing unclassified vs classified reads removed
+
 ## [v0.0.1]
 
 Initial release